ai summarizer/Backend/requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
python-dotenv==0.19.2
requests==2.28.1
python-multipart==0.0.5
openai==0.27.0
ffmpeg-python==0.2.0
gunicorn==20.1.1
